refactor(watch): extract pass, overlay, placement, and event loop modules
